requirements.txt
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2
django-bootstrap-v5==1.0.11
django-bootstrap5==22.2
django-filter==24.2
djangorestframework==3.15.2
flake8==6.0.0
flake8-isort==6.0.0
geographiclib==2.0
geopy==2.4.1
idna==3.7
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
packaging==24.1
pluggy==1.5.0
py==1.11.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==8.2.2
pytest-django==4.8.0
requests==2.32.3
soupsieve==2.5
sqlparse==0.5.1
tomli==2.0.1
urllib3==2.2.2
